mineru-proxy/requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
